fix: rebase on main, revocation_status null handling, type annotations, WASM validation
- Rebase onto origin/main (includes PR #1417 welcome delivery)
- TS: fix revocation_status ternary to treat null/undefined as "active"
  not "revoked"
- Python: fix isinstance check — dict means revoked, else active
- Python: verified_at float → int (Unix seconds are integers)
- Swift: verifiedAt Double → UInt64 (matches Rust u64)
- Cache doc: correct capacity comment from 1024 to 10000
- scp-core validate_structure: add expires_at > issued_at check
- WASM validate_structure: add ID recomputation, revoked_by == issuer,
  and proof-variant-matches-method checks
- WASM compute_attestation_canonical_bytes: replace unwrap_or with
  error propagation via attestation_required_str helper
- WASM: replace all inline capacity checks with check_registry_capacity

+++ bindings/swift/Sources/SCP/Identity.swift
@@ -604,7 +604,7 @@ public struct IdentityAttestation: Sendable {
     public let verificationMethod: String
 
     /// Unix timestamp (seconds) when the evidence was last verified.
-    public let verifiedAt: Int
+    public let verifiedAt: UInt64
 
     /// Revocation status.
     public let revocationStatus: RevocationStatus
@@ -625,7 +625,7 @@ public struct IdentityAttestation: Sendable {
         platform: String,
         platformHandle: String,
         verificationMethod: String,
-        verifiedAt: Int,
+        verifiedAt: UInt64,
         revocationStatus: RevocationStatus = .active,
         platformId: String? = nil,
         rawJson: String? = nil
